Move constants to the app folder
[noissue]
